ALS-1329 - working on nearest region selection
