+++ app/src/main/res/values-v30/themes.xml
@@ -7,9 +7,5 @@
         <item name="colorOnPrimary">@color/white</item>
         <item name="android:windowLightStatusBar">true</item>
         <item name="android:forceDarkAllowed" tools:targetApi="q">false</item>
-        <item name="android:windowDisablePreview">true</item>
-    </style>
-    <style name="AppTheme.Splash" parent="Theme.MaterialComponents.Light.NoActionBar">
-        <item name="android:windowBackground">@color/white</item>
     </style>
 </resources>
