+++ app/src/main/res/values/themes.xml
@@ -8,9 +8,5 @@
         <!-- Status bar color. -->
         <item name="android:windowLightStatusBar">true</item>
         <item name="android:forceDarkAllowed" tools:targetApi="q">false</item>
-        <item name="android:windowDisablePreview">true</item>
-    </style>
-    <style name="AppTheme.Splash" parent="Theme.MaterialComponents.Light.NoActionBar">
-        <item name="android:windowBackground">@color/white</item>
     </style>
 </resources>
